Build 3.1.3..
Department wird kopiert

+++ src/main/java/de/infomotion/kw/demo/model/kwdb/Department.java
@@ -21,5 +21,5 @@ public class Department {
 	private String departmentName;
 
 	@Column(name = "kostenstelle", precision = 38, nullable = true)
-	private Integer costCenter;
+	private BigInteger costCenter;
 }
